ENH: user config, lock files, brain extraction
Three changes in this commit:

User config: Moving all user specified options and paths to a dedicated text file, "user_config.sh". The main script sources this config file- For example, the brain extraction method is set here, as are the TODO/DONE steps.
Removal of lock files: These are not intuitive and I think cause more trouble than they're worth. I've updated the steps to instead check for the output of the step. It runs if the output is not found (thus serving as the lock file). The step directory can be removed if it needs to be re-processed. I plan to add an additional script which clears output files for the specified steps.
Step 4 brain extraction: I created a docker container for Davood's dmri3d- this should now run from any setup. Also some small adjustments to the step to organize the segmentation outputs and prevent junk files from being sent through the segmentation tool.